requirements.txt
msal
requests
pandas
azure-identity
azure-keyvault-secrets
pyodbc
sqlalchemy
python-dotenv
